requirements.txt
streamlit==1.28.1
pdfplumber==0.10.3
python-docx==1.1.0
pandas==2.1.4
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
requests==2.31.0 